api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1
python-multipart==0.0.20
aiofiles==24.1.0
httpx==0.28.1
openai==1.58.1
google-generativeai==0.8.4
pytest==8.3.4
pytest-asyncio==0.25.0
